+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 1.0.0
+
+* **BREAKING CHANGE:** Migrate to [FFmpegKit for Flutter](https://pub.dev/packages/ffmpeg_kit_flutter).
+  `saveTrimmedVideo()` method is not async now, you'll need to use the callback `onSave: (outputPath) {}` to get the trimmed video output path.
+* Add playback timestamp in the `showDuration`
+* Simply configuration
+* Update the plugin versions
+* Update Docs
+
 ## 0.6.0
 
 * Update the plugin versions